Restructure directories to avoid React error
