+++ README.md
@@ -1,3 +1,3 @@
 This repository contains the core user interface components for [vizhub.com](https://vizhub.com).
 
-For local development, run the [test app](./testApp).
+For local development there is a testing app, bootstrapped with [Create React App](https://github.com/facebookincubator/create-react-app).
