+++ package.json
@@ -5,7 +5,8 @@
   "main": "dist/index.js",
   "scripts": {
     "build": "rollup -c",
-    "postinstall": "npm run build"
+    "postinstall": "npm run build",
+    "start": "react-scripts start"
   },
   "repository": {
     "type": "git",
@@ -23,7 +24,8 @@
   "dependencies": {
     "react": "^16.4.1",
     "react-dom": "^16.4.1",
-    "react-measure": "^2.0.2"
+    "react-measure": "^2.0.2",
+    "react-scripts": "1.1.4"
   },
   "author": "Datavis Tech INC.",
   "license": "GPL-3.0",
