+++ src/index.js
@@ -1,1 +1,6 @@
-export { VisualizationRunner } from './visualizationRunner/index';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './index.css';
+import App from './App';
+
+ReactDOM.render(<App />, document.getElementById('root'));
