+++ src/visualizationRunner/computeScale.js
@@ -1,3 +1,4 @@
+// Shrink if necessary. Do not grow.
 export const computeScale = (boundsWidth, width) => (
   boundsWidth > width
     ? 1
