--- testApp/README.md
@@ -1,3 +1,0 @@
-This directory contains a testing app, which provides an environment for developing.
-
-Bootstrapped with [Create React App](https://github.com/facebookincubator/create-react-app).
